+++ crates/feagi-services/src/types/dtos.rs
@@ -151,21 +151,21 @@ pub struct CorticalAreaInfo {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub encoding_format: Option<String>,
 
-    /// Cortical unit index (`CorticalUnitIndex`): byte 7 of the 8-byte cortical ID —
-    /// which instance of this I/O cortical unit type (aligns with motor/sensory grouping
-    /// on the wire and with Brain Visualizer `unit_id`).
+    /// Cortical unit index (`CorticalUnitIndex`): little-endian u16 in bytes 6-7
+    /// of the 8-byte cortical ID — which instance of this I/O cortical unit type
+    /// (aligns with motor/sensory grouping on the wire and with Brain Visualizer `unit_id`).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub unit_id: Option<u8>,
+    pub unit_id: Option<u16>,
 
-    /// Cortical sub-unit index (`CorticalSubUnitIndex`): byte 6 — which cortical area
-    /// inside a multi-area unit (e.g. vision segments). Matches Brain Visualizer `subunit_id`.
+    /// Cortical sub-unit index (`CorticalSubUnitIndex`): flag bits 4-7 — which cortical
+    /// area inside a multi-area unit (e.g. vision segments). Matches Brain Visualizer `subunit_id`.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub subunit_id: Option<u8>,
 
-    /// Same as [`Self::unit_id`] (CorticalUnitIndex, byte 7). Kept for clients that
+    /// Same as [`Self::unit_id`] (CorticalUnitIndex, bytes 6-7). Kept for clients that
     /// already read `group_id` from older API responses.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub group_id: Option<u8>,
+    pub group_id: Option<u16>,
 
     /// IO coding signage (e.g., "Percentage Signed", "Percentage Unsigned")
     #[serde(skip_serializing_if = "Option::is_none")]
